requirements.txt
fastapi
uvicorn
requests
pandas
spacy
networkx
torch
ollama
python-dotenv